+++ src/deviceDetailsChannelTableModel.cpp
@@ -262,27 +262,43 @@ void DeviceDetailsChannelTableModelPrivate::resetChangedData()
 void DeviceDetailsChannelTableModelPrivate::channelConnectionsUpdate(la::avdecc::UniqueIdentifier const& entityId)
 {
 	Q_Q(DeviceDetailsChannelTableModel);
-	int row = 0;
+	auto row = 0;
+	auto updateModel = false;
 	for (auto& node : _nodes)
 	{
 		auto& channelConnectionManager = avdecc::ChannelConnectionManager::getInstance();
 		if (node.connectionInformation->sourceClusterChannelInfo->direction == avdecc::ChannelConnectionDirection::OutputToInput)
 		{
+			// update the node connection information (forward)
 			node.connectionInformation = channelConnectionManager.getChannelConnections(node.connectionInformation->sourceEntityId, *node.connectionInformation->sourceClusterChannelInfo);
-			auto begin = q->index(0, static_cast<int>(DeviceDetailsChannelTableModelColumn::Connection), QModelIndex());
-			auto end = q->index(static_cast<int>(_nodes.size()), static_cast<int>(DeviceDetailsChannelTableModelColumn::ConnectionStatus), QModelIndex());
-			q->dataChanged(begin, end, QVector<int>(Qt::DisplayRole));
+			updateModel = true;
 		}
 		else
 		{
-			if (node.connectionInformation->sourceEntityId == entityId)
+			// if either the connectionInfo source or one of the targets is the entity that the call refers to...
+			if (node.connectionInformation->sourceEntityId == entityId
+					|| std::find_if(node.connectionInformation->targets.begin(), node.connectionInformation->targets.end(),
+							 [entityId](std::shared_ptr<avdecc::TargetConnectionInformation>& targetInfo)
+							 {
+								 return (targetInfo->targetEntityId == entityId);
+							 })
+							 != node.connectionInformation->targets.end())
 			{
+				// ...update the node connection information (reverse)
 				node.connectionInformation = channelConnectionManager.getChannelConnectionsReverse(node.connectionInformation->sourceEntityId, *node.connectionInformation->sourceClusterChannelInfo);
-				auto indexConnection = q->index(row, static_cast<int>(DeviceDetailsChannelTableModelColumn::Connection), QModelIndex());
-				q->dataChanged(indexConnection, indexConnection, QVector<int>(Qt::DisplayRole));
-				auto indexConnectionStatus = q->index(row, static_cast<int>(DeviceDetailsChannelTableModelColumn::ConnectionStatus), QModelIndex());
-				q->dataChanged(indexConnectionStatus, indexConnectionStatus, QVector<int>(Qt::DisplayRole));
+				updateModel = true;
 			}
+		}
+
+		if (updateModel)
+		{
+			// trigger updating the model
+			auto indexConnection = q->index(row, static_cast<int>(DeviceDetailsChannelTableModelColumn::Connection), QModelIndex());
+			if (indexConnection.isValid())
+				q->dataChanged(indexConnection, indexConnection, { Qt::DisplayRole });
+			auto indexConnectionStatus = q->index(row, static_cast<int>(DeviceDetailsChannelTableModelColumn::ConnectionStatus), QModelIndex());
+			if (indexConnectionStatus.isValid())
+				q->dataChanged(indexConnectionStatus, indexConnectionStatus, { Qt::DisplayRole });
 		}
 
 		row++;
@@ -295,7 +311,8 @@ void DeviceDetailsChannelTableModelPrivate::channelConnectionsUpdate(la::avdecc:
 void DeviceDetailsChannelTableModelPrivate::channelConnectionsUpdate(std::set<std::pair<la::avdecc::UniqueIdentifier, avdecc::ChannelIdentification>> channels)
 {
 	Q_Q(DeviceDetailsChannelTableModel);
-	int row = 0;
+	auto row = 0;
+	auto updateModel = false;
 	for (auto& node : _nodes)
 	{
 		auto& channelConnectionManager = avdecc::ChannelConnectionManager::getInstance();
@@ -304,21 +321,27 @@ void DeviceDetailsChannelTableModelPrivate::channelConnectionsUpdate(std::set<st
 			// in the case of the talker we can't know which channels have to be updated without getting the changes from the ChannelConnectionManager
 			// therefor all talkers are updated for now. To improve on this talker connections would need to be cached in the ChannelConnectionManager like the listenerChannelConnections are cached.
 			node.connectionInformation = channelConnectionManager.getChannelConnections(node.connectionInformation->sourceEntityId, *node.connectionInformation->sourceClusterChannelInfo);
-			auto indexConnection = q->index(row, static_cast<int>(DeviceDetailsChannelTableModelColumn::Connection), QModelIndex());
-			q->dataChanged(indexConnection, indexConnection, QVector<int>(Qt::DisplayRole));
-			auto indexConnectionStatus = q->index(row, static_cast<int>(DeviceDetailsChannelTableModelColumn::ConnectionStatus), QModelIndex());
-			q->dataChanged(indexConnectionStatus, indexConnectionStatus, QVector<int>(Qt::DisplayRole));
+			updateModel = true;
 		}
 		else
 		{
 			if (channels.find(std::make_pair(node.connectionInformation->sourceEntityId, *node.connectionInformation->sourceClusterChannelInfo)) != channels.end())
 			{
+				// update the node connection information (reverse)
 				node.connectionInformation = channelConnectionManager.getChannelConnectionsReverse(node.connectionInformation->sourceEntityId, *node.connectionInformation->sourceClusterChannelInfo);
-				auto indexConnection = q->index(row, static_cast<int>(DeviceDetailsChannelTableModelColumn::Connection), QModelIndex());
-				q->dataChanged(indexConnection, indexConnection, QVector<int>(Qt::DisplayRole));
-				auto indexConnectionStatus = q->index(row, static_cast<int>(DeviceDetailsChannelTableModelColumn::ConnectionStatus), QModelIndex());
-				q->dataChanged(indexConnectionStatus, indexConnectionStatus, QVector<int>(Qt::DisplayRole));
+				updateModel = true;
 			}
+		}
+
+		if (updateModel)
+		{
+			// trigger updating the model
+			auto indexConnection = q->index(row, static_cast<int>(DeviceDetailsChannelTableModelColumn::Connection), QModelIndex());
+			if (indexConnection.isValid())
+				q->dataChanged(indexConnection, indexConnection, { Qt::DisplayRole });
+			auto indexConnectionStatus = q->index(row, static_cast<int>(DeviceDetailsChannelTableModelColumn::ConnectionStatus), QModelIndex());
+			if (indexConnectionStatus.isValid())
+				q->dataChanged(indexConnectionStatus, indexConnectionStatus, { Qt::DisplayRole });
 		}
 
 		row++;
@@ -339,7 +362,8 @@ void DeviceDetailsChannelTableModelPrivate::updateAudioClusterName(la::avdecc::U
 			if (entityID == node.connectionInformation->sourceEntityId && configurationIndex == node.connectionInformation->sourceClusterChannelInfo->configurationIndex && audioClusterIndex == node.connectionInformation->sourceClusterChannelInfo->clusterIndex)
 			{
 				auto indexConnectionStatus = q->index(row, static_cast<int>(DeviceDetailsChannelTableModelColumn::ChannelName), QModelIndex());
-				q->dataChanged(indexConnectionStatus, indexConnectionStatus, QVector<int>(Qt::DisplayRole));
+				if (indexConnectionStatus.isValid())
+					q->dataChanged(indexConnectionStatus, indexConnectionStatus, QVector<int>(Qt::DisplayRole));
 			}
 		}
 		row++;