+++ src/mediaClock/domainTreeModel.cpp
@@ -714,7 +714,7 @@ bool DomainTreeModelPrivate::setData(QModelIndex const& index, QVariant const& v
 		if (treeItem->type() == AbstractTreeItem::Domain)
 		{
 			auto* domainTreeItem = static_cast<DomainTreeItem*>(treeItem);
-			if (*domainTreeItem->domainSamplingRate().first != value.toUInt())
+			if (domainTreeItem->domainSamplingRate().first && *domainTreeItem->domainSamplingRate().first != value.toUInt())
 			{
 				domainTreeItem->setDomainSamplingRate(la::avdecc::entity::model::SamplingRate(value.toUInt()));
 				emit q->domainSetupChanged();
@@ -1476,7 +1476,10 @@ void SampleRateDomainDelegate::paint(QPainter* painter, QStyleOptionViewItem con
 		editor.getLabel()->setText(elidedText);
 		for (auto const& sampleRate : sampleRates)
 		{
-			editor.getComboBox()->addItem(sampleRate.second, sampleRate.first->getValue());
+			if (sampleRate.first)
+			{
+				editor.getComboBox()->addItem(sampleRate.second, sampleRate.first->getValue());
+			}
 		}
 		if (selectedSampleRate.first)
 		{